sped-backend/requirements.txt
anyio==4.0.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
black==23.9.1
boto3==1.34.74
bs4==0.0.2
certifi~=2024.2.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
faiss-cpu==1.8.0
fastapi==0.110.0
fastapi-versioning==0.10.0
flake8==6.1.0
flask==3.0.2
flask-cors==4.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
networkx==3.3
nltk~=3.8.1
openpyxl==3.1.2
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.7.0
py-mon==2.0.5
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.6.4
pyflakes==3.1.0
PyJWT==2.8.0
pymongo~=4.6.2
pytest==7.4.2
python-dotenv==1.0.1
python-jose==3.3.0
rapidocr-onnxruntime==1.3.15
requests~=2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
spacy==3.7.4
starlette==0.36.3
tenacity==8.3.0
typing_extensions==4.11.0rc1
uvicorn==0.30.1
xlrd==2.0.1
langchain==0.2.5
langchain_openai==0.1.9
